Fixes for re-enabling more MSVC level 4 warnings: chrome/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Assignment inside conditional
* Constant value was truncated
* Possibly-uninitialized local variable

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/377683002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283305 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9fc5ea7f00b59473330eb52550059da2dcf78b61

+++ chrome/installer/setup/setup_util.cc
@@ -242,7 +242,7 @@ bool DeleteFileFromTempProcess(const base::FilePath& path,
       }
     } else {
       PLOG(ERROR) << "VirtualAllocEx";
-      ::TerminateProcess(pi.hProcess, ~0);
+      ::TerminateProcess(pi.hProcess, ~static_cast<UINT>(0));
     }
     ::CloseHandle(pi.hThread);
     ::CloseHandle(pi.hProcess);
